Move utils::testing to test_utils
Have a single module with test utils instead of scattered functions.
