+++ src/utils.rs
@@ -125,36 +125,3 @@ where
 
     Ok(())
 }
-
-#[cfg(test)]
-pub mod testing {
-    use crate::config::SerialisableCertificate;
-    use crate::dirs::{Dirs, OverRide};
-    use crate::NodeInfo;
-    use rand::Rng;
-    use std::env;
-    use std::net::{Ipv4Addr, SocketAddr, SocketAddrV4};
-    use std::path::PathBuf;
-
-    pub fn test_dirs() -> Dirs {
-        Dirs::Overide(OverRide::new(&unwrap!(tmp_rand_dir().to_str())))
-    }
-
-    pub fn rand_node_info() -> NodeInfo {
-        let peer_cert_der = SerialisableCertificate::default().cert_der;
-        let mut rng = rand::thread_rng();
-        let port: u16 = rng.gen();
-        let peer_addr = SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, port));
-        NodeInfo {
-            peer_addr,
-            peer_cert_der,
-        }
-    }
-
-    fn tmp_rand_dir() -> PathBuf {
-        let fname = format!("quic_p2p_tests_{:016x}", rand::random::<u64>());
-        let mut path = env::temp_dir();
-        path.push(fname);
-        path
-    }
-}
